Add anomaly tests to CategoriesOperations.test and GroupCategoriesOperations.test
